% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchSapSaTA}
\alias{launchSapSaTA}
\title{Launch SapSaTA}
\usage{
launchSapSaTA(...)
}
\value{
Shiny application.
}
\description{
Launch SapSaTA
}
\details{
Launch SapSaTA
}
\keyword{Homo}
\keyword{Human}
\keyword{Sapien}
\keyword{atlas}
\keyword{sc-RNASeq}
\keyword{spatialomics}
\keyword{spermatogenesis}
\keyword{spermiogenesis}
\keyword{testes}
\keyword{testis}
